Move exception extended formatting to its own class
